requirements.txt
mutagen==1.41.1
pygame==1.9.4
pydub==0.23.0
PyAudio==0.2.11
pandas==0.23.0
tflearn==0.3.2
eyeD3==0.8.7
numpy==1.14.3
Pillow==5.3.0
PyQt5==5.11.3
